Refactor version package to align with new application package
