+++ application/application_test.go
@@ -5,10 +5,10 @@ import (
 	. "github.com/onsi/gomega"
 
 	"github.com/tidepool-org/platform/application"
+	"github.com/tidepool-org/platform/application/version"
+	_ "github.com/tidepool-org/platform/application/version/test"
 	"github.com/tidepool-org/platform/config"
 	"github.com/tidepool-org/platform/config/env"
-	"github.com/tidepool-org/platform/version"
-	_ "github.com/tidepool-org/platform/version/test"
 )
 
 var _ = Describe("Application", func() {
