+++ application/version/reporter_test.go
@@ -4,16 +4,16 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	"github.com/tidepool-org/platform/version"
+	"github.com/tidepool-org/platform/application/version"
 )
 
-var _ = Describe("Default", func() {
-	Context("NewDefaultReporter", func() {
+var _ = Describe("Reporter", func() {
+	Context("NewReporter", func() {
 		It("returns successfully", func() {
 			version.Base = "1.2.3"
 			version.ShortCommit = "4567890"
 			version.FullCommit = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmn"
-			Expect(version.NewDefaultReporter()).ToNot(BeNil())
+			Expect(version.NewReporter()).ToNot(BeNil())
 		})
 	})
 })
