+++ tool/mongo/tool_test.go
@@ -4,11 +4,11 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
+	"github.com/tidepool-org/platform/application/version"
+	_ "github.com/tidepool-org/platform/application/version/test"
 	"github.com/tidepool-org/platform/config"
 	"github.com/tidepool-org/platform/config/env"
 	"github.com/tidepool-org/platform/tool/mongo"
-	"github.com/tidepool-org/platform/version"
-	_ "github.com/tidepool-org/platform/version/test"
 )
 
 var _ = Describe("Tool", func() {
